WIP: try to port to using lmdb-zero
This doesn't build; I couldn't figure out how to get the lifetimes
to work. lmdb-zero uses an Accessor type to mediate access to data,
and pointers to keys and values in the DB have the same lifetime as
this accessor type (unlike lmdb-rs where the lifetimes are tied to the
transaction itself). I think maybe this permits lmdb-zero to safely read
and write data in the same transaction. From the docs:

> One unusual property of this crate are the ConstAccessor and
> WriteAccessor structs, which are obtained once from a transaction
> and used to perform actual data manipulation. These are needed to
> work with the borrow checker: Cursors have a lifetime bound by their
> transaction and thus borrow it, so we need something else to permit
> borrowing mutable data. The accessors reflect this borrowing: Reading
> from the database requires an immutable borrow of the accessor, while
> writing (which may invalidate pointers) requires a mutable borrow of
> the accessor, thus causing the borrow checker to ensure that all read
> accesses are disposed before any write.

But this also has the effect that ElementTable can't return keys or vals
from the get() or iter() methods, at least without a lot of refactoring
(to store an Accessor in the table itself). And Accessors are !Send so
storing one on the table means you can't implement par_iter().